properly split and register IsbrhTessellatorAbuseTransformer
