+++ src/main/java/com/gtnewhorizons/angelica/loading/rfb/transformers/RFBIsbrhTessellatorAbuseTransformer.java
@@ -1,7 +1,7 @@
 package com.gtnewhorizons.angelica.loading.rfb.transformers;
 
-import com.gtnewhorizons.angelica.loading.fml.transformers.IsbrhTessellatorAbuseClassTransformer;
 import com.gtnewhorizons.angelica.loading.shared.AngelicaClassDump;
+import com.gtnewhorizons.angelica.loading.shared.transformers.IsbrhTessellatorAbuseTransform;
 import com.gtnewhorizons.retrofuturabootstrap.api.ClassHeaderMetadata;
 import com.gtnewhorizons.retrofuturabootstrap.api.ClassNodeHandle;
 import com.gtnewhorizons.retrofuturabootstrap.api.ExtensibleClassLoader;
@@ -12,12 +12,15 @@ import org.jetbrains.annotations.Nullable;
 
 import java.util.jar.Manifest;
 
-/** RfbClassTransformer wrapper for {@link IsbrhTessellatorAbuseClassTransformer} */
+/** RfbClassTransformer wrapper for {@link IsbrhTessellatorAbuseTransform} */
 public class RFBIsbrhTessellatorAbuseTransformer implements RfbClassTransformer {
-    private final IsbrhTessellatorAbuseClassTransformer inner;
+
+    private final IsbrhTessellatorAbuseTransform inner;
+    private final boolean isObf;
 
     public RFBIsbrhTessellatorAbuseTransformer(boolean isObf) {
-        inner = new IsbrhTessellatorAbuseClassTransformer(isObf);
+        this.inner = new IsbrhTessellatorAbuseTransform();
+        this.isObf = isObf;
     }
 
     @Pattern("[a-z0-9-]+")
@@ -42,13 +45,13 @@ public class RFBIsbrhTessellatorAbuseTransformer implements RfbClassTransformer 
         @NotNull ClassNodeHandle classNode) {
         if (!classNode.isPresent()) return false;
         final ClassHeaderMetadata metadata = classNode.getOriginalMetadata();
-        return metadata != null && metadata.binaryInterfaceNames.contains(IsbrhTessellatorAbuseClassTransformer.ISBRH);
+        return metadata != null && metadata.binaryInterfaceNames.contains(IsbrhTessellatorAbuseTransform.ISBRH);
     }
 
     @Override
     public void transformClass(@NotNull ExtensibleClassLoader classLoader, @NotNull RfbClassTransformer.Context context,
         @Nullable Manifest manifest, @NotNull String className, @NotNull ClassNodeHandle classNode) {
-        if (inner.transformClassNode(className, classNode.getNode())) {
+        if (inner.transformClassNode(classNode.getNode(), this.isObf)) {
             classNode.computeMaxs();
             AngelicaClassDump.dumpRFBClass(className, classNode, this);
         }
